pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "slidecat"
version = "0.1.0"
description = "CLI tool to split and merge PowerPoint files"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Kazunori Otani", email = "katzchang@gmail.com"}
]
dependencies = [
    "python-pptx>=0.6.21",
    "click>=8.0.0",
]

[project.scripts]
slidecat = "slidecat.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]
